+++ spec/socket.spec.ts
@@ -55,7 +55,6 @@ test('socket() connects with various URLs', t => {
     t.ok(await socket(`tcp:127.0.0.1:${port}`)(), 'prot:ipv4:port')
     t.ok(await socket(`tcp://${host}:${port}`)(), 'prot://host:port')
     t.ok(await socket(`tcp://127.0.0.1:${port}`)(), 'prot://ipv4:port')
-    t.ok(await socket(`${host}:${port}`)(), 'host:port')
     t.ok(await socket(`http://${host}:${port}`)(), 'http://host:port')
     server.close()
     t.end()